+++ src/constants/IdStructure.ts
@@ -1,0 +1,4 @@
+export enum IdStructure {
+    SUBPATH = 'subpath',
+    DELIMITER = 'delimiter'
+}