+++ src/constants/index.ts
@@ -2,3 +2,4 @@ export * from './LanguageCultures';
 export * from './I18nPrefix';
 export * from './I18nDelimiter';
 export * from './ReferenceBehavior';
+export * from './IdStructure';