+++ src/types/ITranslationRef.ts
@@ -1,5 +1,9 @@
 export interface ITranslationRef {
-    _type: 'reference';
-    _ref: string;
-    _weak: boolean;
+    _key: string;
+    lang: string;
+    ref: {
+        _type: 'reference';
+        _ref: string;
+        _weak: boolean;
+    };
 }