+++ src/types/Ti18nConfig.ts
@@ -1,11 +1,12 @@
 import { TLanguagesOption } from './TLanguagesOption';
 import { TMessagesConfig } from './TMessagesConfig';
 import { TFieldNamesConfig } from './TFieldNamesConfig';
-import { ReferenceBehavior } from '../constants';
+import { IdStructure, ReferenceBehavior } from '../constants';
 
 export type Ti18nConfig = {
   base?: string;
   languages?: TLanguagesOption;
+  idStructure?: IdStructure;
   referenceBehavior?: ReferenceBehavior;
   fieldNames?: TFieldNamesConfig;
   messages?: TMessagesConfig;