+++ src/utils/buildDocId.ts
@@ -1,3 +1,8 @@
-import { I18nPrefix } from "../constants";
+import { I18nDelimiter, I18nPrefix, IdStructure } from '../constants';
+import { getConfig } from './getConfig';
 
-export const buildDocId = (id: string, lang: string | null) => `${I18nPrefix}.${id}.${lang || '*'}`;
+export const buildDocId = (id: string, lang: string | null) => {
+    const config = getConfig();
+    if (config.idStructure === IdStructure.DELIMITER) return `${id}${I18nDelimiter}${lang || '*'}`
+    return `${I18nPrefix}.${id}.${lang || '*'}`;
+}