+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v5.0.4
 * Added `referenceBehavior` setting (allowing users to choose whether they want references from the parent document, to the translationss)
-* Added `idStructure` setting (allowing users to choose between subpath IDs and delimiter IDss)
+* Added `idStructure` setting (allowing users to choose between subpath IDs and delimiter IDs)
+* Fixed an issue where nested object types did not support the i18n field.
 
 ## v5.0.3
 * Updated document list filter to include root drafts