+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG
 
 ## v5.0.4
-* Added `referenceBehavior` setting (allowing users to choose whether they want references from the parent document, to the translationss)
+* [#16](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/16) Added `referenceBehavior` setting (allowing users to choose whether they want references from the parent document, to the translations)
+* [#17](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/17) Fixed an issue where nested object types did not support the i18n field.
+* [#15](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/15) Fixed missing sorting menu
 * Added `idStructure` setting (allowing users to choose between subpath IDs and delimiter IDs)
-* Fixed an issue where nested object types did not support the i18n field.
 
 ## v5.0.3
 * Updated document list filter to include root drafts