+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## v5.0.5
+* Fix `referenceBehavior` not accepting the user config
+
 ## v5.0.4
 * [#16](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/16) Added `referenceBehavior` setting (allowing users to choose whether they want references from the parent document, to the translations)
 * [#17](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/17) Fixed an issue where nested object types did not support the i18n field.